Amesos2: Made a call to max explicit
The compiler was unable to resolve an ambiguity coming from Sacado.
I find that making these things explicit helps fix these issues.

+++ packages/amesos2/src/Amesos2_SolverCore_def.hpp
@@ -368,7 +368,7 @@ SolverCore<ConcreteSolver,Matrix,Vector>::describe(
   for( size_t dec = 10; dec < globalNumRows_; dec *= 10 ) {
     ++width;
   }
-  width = std::max<size_t>(width,11) + 2;
+  width = std::max<size_t>(width,size_t(11)) + 2;
   Teuchos::OSTab tab(out);
   //    none: print nothing
   //     low: print O(1) info from node 0
